feat(streaming): migrate AppendOnlyTopNExecutor to ExecutorV2
