+++ src/stream/src/executor_v2/top_n_executor.rs
@@ -1,10 +1,17 @@
+use std::sync::Arc;
+
 use async_trait::async_trait;
 use futures::StreamExt;
 use futures_async_stream::try_stream;
-use risingwave_common::array::StreamChunk;
+use itertools::Itertools;
+use risingwave_common::array::column::Column;
+use risingwave_common::array::{Op, Row, StreamChunk};
 use risingwave_common::catalog::Schema;
+use risingwave_common::util::chunk_coalesce::DataChunkBuilder;
 
-use crate::executor_v2::error::{StreamExecutorResult, TracedStreamExecutorError};
+use crate::executor_v2::error::{
+    StreamExecutorError, StreamExecutorResult, TracedStreamExecutorError,
+};
 use crate::executor_v2::{BoxedExecutor, BoxedMessageStream, Executor, Message, PkIndicesRef};
 
 #[async_trait]
@@ -96,5 +103,35 @@ where
                 }
             };
         }
+    }
+}
+
+pub(crate) fn generate_output(
+    new_rows: Vec<Row>,
+    new_ops: Vec<Op>,
+    schema: &Schema,
+) -> StreamExecutorResult<StreamChunk> {
+    if !new_rows.is_empty() {
+        let mut data_chunk_builder = DataChunkBuilder::new_with_default_size(schema.data_types());
+        for row in &new_rows {
+            data_chunk_builder
+                .append_one_row_ref(row.into())
+                .map_err(StreamExecutorError::eval_error)?;
+        }
+        // since `new_rows` is not empty, we unwrap directly
+        let new_data_chunk = data_chunk_builder
+            .consume_all()
+            .map_err(StreamExecutorError::eval_error)?
+            .unwrap();
+        let new_stream_chunk = StreamChunk::new(new_ops, new_data_chunk.columns().to_vec(), None);
+        Ok(new_stream_chunk)
+    } else {
+        let columns = schema
+            .create_array_builders(0)
+            .unwrap()
+            .into_iter()
+            .map(|x| Column::new(Arc::new(x.finish().unwrap())))
+            .collect_vec();
+        Ok(StreamChunk::new(vec![], columns, None))
     }
 }
